esearch

search NCBI db using esearch API. retun

esearch(
  query,
  db = "pubmed",
  api_key = Sys.getenv("NCBI_API_KEY"),
  usehistory = TRUE,
  retmax = 20,
  retstart = 0,
  rettype = "uilist"
)

Arguments

query

NCBI string query. for more information please read

db

NCBI DBs described in details

api_key

NCBI api_key. you can get your own api_key from NCBI setting page after log-in. To use the function easily, you need to declare "NCBI_API_KEY" in .Renviron file.

usehistory

T/F. save result in your search history. you can also get WebEnv information. (default=T)

retmax

total number of unique identifiers from the retrived set to be shown in the output (default=20)

retstart

number of skipping from the search (Default=0)

rettype

(optional) "uilist" or "count" (default="uilist"). if the value is count, the function only redirects the total number of query results.

datetype

(optional) one from c("crdt","edat","pdat","mhda")

Value

list or vector object.

Details

available DBs :

https://www.ncbi.nlm.nih.gov/books/NBK25497/table/chapter2.T._entrez_unique_identifiers_ui/?report=objectonly

or use the function einfo()

Followings are examples of search field tags from https://pubmed.ncbi.nlm.nih.gov/help/#proximity-searching. You may also get the information by einfo(db="pubmed")$Field

Affiliation ad	All Fields all	Article Identifier aid

Author au	Author Identifier auid	Last Author Name lastau

PMID pmid	Title/Abstract tiab	Grant Number gr

Publication Date dp	Title ti	MeSH Terms mh

Text Words tw	Title ti	MeSH Major Topic majr